// app/types/Player.tsx
export interface Player {
  id: string;
  name: string;
  image?: string;
  shirt?: string | number;
  state?: string;
  playerPosition?: string;
  age?: number;
  playerHeight?: string;
  games?: number;
  position?: string;
  cleanSheet?: number;
  peneltyStop?: number;
  goals?: number;
  assists?: number;
}